fix: injectCitationLinks 移出 ChatPage，并补上被 if 掩盖的断言
CI 的前端 lint 跑的是 `--max-warnings 0`（不是 CLAUDE.md 写的 50），为可测试
而 export injectCitationLinks 触发 react-refresh/only-export-components，直接
卡住构建。按该规则本意把函数与 CITATION_URL_SCHEME 拆到 utils/citationLinks.ts，
测试随之移到同目录。

同时修掉测试里一处空转断言：原用例把结果包在 `if (chunk !== "-1")` 里，退化时
整条静默通过。改成硬断言后立刻暴露它从未真正跑过所测路径——引文「無學身語業」
只有 5 字，而 extractPrecedingQuote 的正则下限是 6，抽取直接返回 null，卷段
同源逻辑压根没触发。换成足够长的引文后断言才真正成立。

夹具补齐后又暴露 pickSourceForQuote 对 chunk_text 无保护：该字段在类型上必填，
但历史消息的 sources 是存回 JSON 再读出的，不经任何运行时校验，缺字段时这里
抛 TypeError 会让整条消息渲染失败。加一行跳过。
